+++ zcash_client_backend/src/data_api.rs
@@ -2,7 +2,7 @@
 
 use std::{
     collections::{BTreeMap, HashMap},
-    fmt::{self, Debug},
+    fmt::Debug,
     io,
     num::{NonZeroU32, TryFromIntError},
 };
@@ -23,7 +23,7 @@ use zcash_primitives::{
         },
         Transaction, TxId,
     },
-    zip32::AccountId,
+    zip32::{AccountId, Scope},
 };
 
 use crate::{
@@ -31,7 +31,7 @@ use crate::{
     decrypt::DecryptedOutput,
     keys::{UnifiedFullViewingKey, UnifiedSpendingKey},
     proto::service::TreeState,
-    wallet::{ReceivedSaplingNote, WalletTransparentOutput, WalletTx},
+    wallet::{NoteId, ReceivedNote, Recipient, WalletTransparentOutput, WalletTx},
 };
 
 use self::chain::CommitmentTreeRoot;
@@ -358,7 +358,7 @@ pub trait SaplingInputSource {
         &self,
         txid: &TxId,
         index: u32,
-    ) -> Result<Option<ReceivedSaplingNote<Self::NoteRef>>, Self::Error>;
+    ) -> Result<Option<ReceivedNote<Self::NoteRef>>, Self::Error>;
 
     /// Returns a list of spendable Sapling notes sufficient to cover the specified target value,
     /// if possible.
@@ -368,7 +368,7 @@ pub trait SaplingInputSource {
         target_value: Amount,
         anchor_height: BlockHeight,
         exclude: &[Self::NoteRef],
-    ) -> Result<Vec<ReceivedSaplingNote<Self::NoteRef>>, Self::Error>;
+    ) -> Result<Vec<ReceivedNote<Self::NoteRef>>, Self::Error>;
 }
 
 /// A trait representing the capability to query a data store for unspent transparent UTXOs
@@ -602,20 +602,20 @@ impl BlockMetadata {
 /// decrypted and extracted from a [`CompactBlock`].
 ///
 /// [`CompactBlock`]: crate::proto::compact_formats::CompactBlock
-pub struct ScannedBlock<Nf> {
+pub struct ScannedBlock<Nf, S> {
     block_height: BlockHeight,
     block_hash: BlockHash,
     block_time: u32,
     sapling_tree_size: u32,
     orchard_tree_size: u32,
-    transactions: Vec<WalletTx<Nf>>,
+    transactions: Vec<WalletTx<Nf, S>>,
     sapling_nullifier_map: Vec<(TxId, u16, Vec<sapling::Nullifier>)>,
     sapling_commitments: Vec<(sapling::Node, Retention<BlockHeight>)>,
     orchard_nullifier_map: Vec<(TxId, u16, Vec<orchard::note::Nullifier>)>,
     orchard_commitments: Vec<(orchard::note::NoteCommitment, Retention<BlockHeight>)>,
 }
 
-impl<Nf> ScannedBlock<Nf> {
+impl<Nf, S> ScannedBlock<Nf, S> {
     /// Constructs a new `ScannedBlock`
     #[allow(clippy::too_many_arguments)]
     pub fn from_parts(
@@ -624,7 +624,7 @@ impl<Nf> ScannedBlock<Nf> {
         block_time: u32,
         sapling_tree_size: u32,
         orchard_tree_size: u32,
-        transactions: Vec<WalletTx<Nf>>,
+        transactions: Vec<WalletTx<Nf, S>>,
         sapling_nullifier_map: Vec<(TxId, u16, Vec<sapling::Nullifier>)>,
         sapling_commitments: Vec<(sapling::Node, Retention<BlockHeight>)>,
         orchard_nullifier_map: Vec<(TxId, u16, Vec<orchard::note::Nullifier>)>,
@@ -670,7 +670,7 @@ impl<Nf> ScannedBlock<Nf> {
     }
 
     /// Returns the list of transactions from the block that are relevant to the wallet.
-    pub fn transactions(&self) -> &[WalletTx<Nf>] {
+    pub fn transactions(&self) -> &[WalletTx<Nf, S>] {
         &self.transactions
     }
 
@@ -754,81 +754,6 @@ pub struct SentTransaction<'a> {
     pub fee_amount: Amount,
     #[cfg(feature = "transparent-inputs")]
     pub utxos_spent: Vec<OutPoint>,
-}
-
-/// A shielded transfer protocol supported by the wallet.
-#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
-pub enum ShieldedProtocol {
-    /// The Sapling protocol
-    Sapling,
-    /// The Orchard protocol
-    Orchard,
-}
-
-/// A unique identifier for a shielded transaction output
-#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
-pub struct NoteId {
-    txid: TxId,
-    protocol: ShieldedProtocol,
-    output_index: u16,
-}
-
-impl NoteId {
-    /// Constructs a new `NoteId` from its parts.
-    pub fn new(txid: TxId, protocol: ShieldedProtocol, output_index: u16) -> Self {
-        Self {
-            txid,
-            protocol,
-            output_index,
-        }
-    }
-
-    /// Returns the ID of the transaction containing this note.
-    pub fn txid(&self) -> &TxId {
-        &self.txid
-    }
-
-    /// Returns the shielded protocol used by this note.
-    pub fn protocol(&self) -> ShieldedProtocol {
-        self.protocol
-    }
-
-    /// Returns the index of this note within its transaction's corresponding list of
-    /// shielded outputs.
-    pub fn output_index(&self) -> u16 {
-        self.output_index
-    }
-}
-
-/// A value pool to which the wallet supports sending transaction outputs.
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub enum PoolType {
-    /// The transparent value pool
-    Transparent,
-    /// A shielded value pool.
-    Shielded(ShieldedProtocol),
-}
-
-impl fmt::Display for PoolType {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            PoolType::Transparent => f.write_str("Transparent"),
-            PoolType::Shielded(ShieldedProtocol::Sapling) => f.write_str("Sapling"),
-            PoolType::Shielded(ShieldedProtocol::Orchard) => f.write_str("Orchard"),
-        }
-    }
-}
-
-/// A type that represents the recipient of a transaction output; a recipient address (and, for
-/// unified addresses, the pool to which the payment is sent) in the case of outgoing output, or an
-/// internal account ID and the pool to which funds were sent in the case of a wallet-internal
-/// output.
-#[derive(Debug, Clone)]
-pub enum Recipient {
-    Transparent(TransparentAddress),
-    Sapling(sapling::PaymentAddress),
-    Unified(UnifiedAddress, PoolType),
-    InternalAccount(AccountId, PoolType),
 }
 
 /// A type that represents an output (either Sapling or transparent) that was sent by the wallet.
@@ -1056,7 +981,7 @@ pub trait WalletWrite: WalletRead {
     /// `blocks` must be sequential, in order of increasing block height
     fn put_blocks(
         &mut self,
-        blocks: Vec<ScannedBlock<sapling::Nullifier>>,
+        blocks: Vec<ScannedBlock<sapling::Nullifier, Scope>>,
     ) -> Result<(), Self::Error>;
 
     /// Updates the wallet's view of the blockchain.
@@ -1143,20 +1068,19 @@ pub mod testing {
         memo::Memo,
         sapling,
         transaction::{components::Amount, Transaction, TxId},
-        zip32::AccountId,
+        zip32::{AccountId, Scope},
     };
 
     use crate::{
         address::{AddressMetadata, UnifiedAddress},
         keys::{UnifiedFullViewingKey, UnifiedSpendingKey},
-        wallet::{ReceivedSaplingNote, WalletTransparentOutput},
+        wallet::{NoteId, ReceivedNote, WalletTransparentOutput},
     };
 
     use super::{
         chain::CommitmentTreeRoot, scanning::ScanRange, AccountBirthday, BlockMetadata,
-        DecryptedTransaction, NoteId, NullifierQuery, SaplingInputSource, ScannedBlock,
-        SentTransaction, WalletCommitmentTrees, WalletRead, WalletSummary, WalletWrite,
-        SAPLING_SHARD_HEIGHT,
+        DecryptedTransaction, NullifierQuery, SaplingInputSource, ScannedBlock, SentTransaction,
+        WalletCommitmentTrees, WalletRead, WalletSummary, WalletWrite, SAPLING_SHARD_HEIGHT,
     };
 
     pub struct MockWalletDb {
@@ -1188,7 +1112,7 @@ pub mod testing {
             &self,
             _txid: &TxId,
             _index: u32,
-        ) -> Result<Option<ReceivedSaplingNote<Self::NoteRef>>, Self::Error> {
+        ) -> Result<Option<ReceivedNote<Self::NoteRef>>, Self::Error> {
             Ok(None)
         }
 
@@ -1198,7 +1122,7 @@ pub mod testing {
             _target_value: Amount,
             _anchor_height: BlockHeight,
             _exclude: &[Self::NoteRef],
-        ) -> Result<Vec<ReceivedSaplingNote<Self::NoteRef>>, Self::Error> {
+        ) -> Result<Vec<ReceivedNote<Self::NoteRef>>, Self::Error> {
             Ok(Vec::new())
         }
     }
@@ -1366,7 +1290,7 @@ pub mod testing {
         #[allow(clippy::type_complexity)]
         fn put_blocks(
             &mut self,
-            _blocks: Vec<ScannedBlock<sapling::Nullifier>>,
+            _blocks: Vec<ScannedBlock<sapling::Nullifier, Scope>>,
         ) -> Result<(), Self::Error> {
             Ok(())
         }