+++ zcash_client_sqlite/src/wallet/init/migrations.rs
@@ -5,6 +5,7 @@ mod addresses_table;
 mod initial_setup;
 mod nullifier_map;
 mod received_notes_nullable_nf;
+mod receiving_key_scopes;
 mod sapling_memo_consistency;
 mod sent_notes_to_internal;
 mod shardtree_support;
@@ -40,17 +41,17 @@ pub(super) fn all_migrations<P: consensus::Parameters + 'static>(
     //                                     |
     //                             v_transactions_net
     //                                     |
-    //                         received_notes_nullable_nf
-    //                        /            |             \
-    //        shardtree_support      nullifier_map       sapling_memo_consistency
-    //                |                                             |
-    //      add_account_birthdays                   v_transactions_transparent_history
-    //                |                                             |
-    // v_sapling_shard_unscanned_ranges               v_tx_outputs_use_legacy_false
-    //                |                                             |
-    //        wallet_summaries                       v_transactions_shielding_balance
-    //                                                              |
-    //                                               v_transactions_note_uniqueness
+    //                                  received_notes_nullable_nf
+    //                                 /            |             \
+    //                 shardtree_support      nullifier_map       sapling_memo_consistency
+    //                  /              \                                      |
+    //      add_account_birthdays   receiving_key_scopes      v_transactions_transparent_history
+    //                  |                                                     |
+    // v_sapling_shard_unscanned_ranges                         v_tx_outputs_use_legacy_false
+    //                  |                                                     |
+    //        wallet_summaries                                 v_transactions_shielding_balance
+    //                                                                        |
+    //                                                          v_transactions_note_uniqueness
     vec![
         Box::new(initial_setup::Migration {}),
         Box::new(utxos_table::Migration {}),
@@ -86,5 +87,8 @@ pub(super) fn all_migrations<P: consensus::Parameters + 'static>(
         Box::new(v_tx_outputs_use_legacy_false::Migration),
         Box::new(v_transactions_shielding_balance::Migration),
         Box::new(v_transactions_note_uniqueness::Migration),
+        Box::new(receiving_key_scopes::Migration {
+            params: params.clone(),
+        }),
     ]
 }